requirements.txt
easydict

numpy
scikit-learn
matplotlib==3.7.1
opencv-python==4.7.0.72
open3d
plyfile
nltk
tqdm
ftfy
regex
transformers

torch_geometric==1.7.2 
torch-cluster==1.6.0    
torch-scatter==2.0.9
torch-sparse==0.6.15

timm==0.4.5